some re-org and cleanup
